Move @StaticInitSafe annotation to the proper package
