+++ core/runtime/src/main/java/io/quarkus/runtime/annotations/StaticInitSafe.java
@@ -1,4 +1,4 @@
-package io.quarkus.runtime.configuration;
+package io.quarkus.runtime.annotations;
 
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
